(*) Isolate thhe index construction process
In order to facilitate investigation of a putative JVM bug, this change makes it  possible to run in isolation for replicating a putative JVM bug.  It should not be merged back to the main branch.

Backporting a few fixes to the merging code that reduce the amount of log noise, and fails faster if the bug manifests.